GR #6626 Stop fetching ivy.xml files eagerly.
The Holy Gradle needs the ivy.xml for a module for two reasons:
* to get the relativePath attribute for dependencies (now off by default);
* to make an offline copy of dependencies with collectDependencies.

So, for most runs these files aren't needed, so there's point us fetching them.
The basic change here is to change the constructor of UnpackModuleVersion to
take a Closure<File> instead of a File and to populate the map of relative
paths lazily.