More fixing of tests.
* custom-gradle-core-plugin\src\test\groovy\holygradle\test\AbstractHolyGradleTest.groovy
** Add in an improvement already made in another branch -- extra logging before
and after each test, so I can see the boundaries in the test output.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\GradleZipHelper.groovy
** It turns out this class was never really tested before, because I always run
with 7-zip at work.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** We shouldn't be visiting children of source dependencies (just as we were
already not visiting source dependencies).  Previuosly we accidentally achieved
this because both parts of the VisitChoice included "&& ivyFileExists" but
source dependencies don't have ivy files, so we never visited their children,
even though "isNewModuleConfiguration" was true for *children* of source deps.
Now that the ivy file is no longer checked (because we're not supporting
explicit relative paths), I need to correct the remaining part of the choice.

* intrepid-plugin\src\test\groovy\holygradle\unpacking\UnpackModuleVersionTest.groovy
** Remove more tests related to the now-removed support for explicit relative
paths.