DefaultPublishPackagesExtensionIntegrationTest passes at last!
* intrepid-plugin\src\test\groovy\holygradle\publishing\DefaultPublishPackagesExtensionIntegrationTest.groovy
** Add before-/afterGenerateDescriptor tasks, and depdencies with other tasks,
because they're needed by other changes.
** Verifying the group, name, and "that we're not skipping link creation" should
all happen before we generate the module descriptor, not wait until we are
actually publishing.  Also, adding the doLast callback for those things can
happen early -- it doesn't need to be in a projectsEvaluated block.
** Run our extra "ivy.xml cleanup" tasks for every publication, not just the
default one created by this class.
** Factor out fixArrowsInXml and likewise do that for every publication.
** New helper method to add (to the default publication) dependencies for all
configurations in the project.  Since 1.5 the ivy-publish plugin no longer does
that automatically.

* intrepid-plugin\src\main\groovy\holygradle\publishing\DefaultPublishPackagesExtension.groovy
** Remove extra logging.