Cleanups to get tests working.  Also, make some new publish behaviour optional.
* custom-gradle-core-plugin\src\init.d\holy-gradle-init.gradle
** Now that I'm actually building without Artifactory, I discovered that a
couple of "hasProperty" checks were always true, because those proerties have
been (for some time) always created, even if set to null.  So test for null
instead.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\ResolvedDependenciesVisitor.groovy
** Comment out temporary logging.

* intrepid-plugin\src\main\groovy\holygradle\publishing\DefaultPublishPackagesExtension.groovy
* intrepid-plugin\src\main\groovy\holygradle\publishing\PublishPackagesExtension.groovy
** Add explicit options to control whether or not the default publication is
created and whether or not to publish private configurations which are not
extended from by other configurations.  (We always have to publish private
configurations which other configurations extend from.)

* intrepid-plugin\src\main\groovy\holygradle\unpacking\PackedDependenciesStateHandler.groovy
** Remove temporary logging.

* intrepid-plugin\src\test\groovy\holygradle\scm\SourceControlRepositoriesTest.groovy
* intrepid-plugin\src\test\groovy\holygradle\scm\SourceControlRepositoriesMultipleSCMFoldersTest.groovy
** Pull the multiple folders test out from being an inner class, because that
doesn't work with the latest Java/Gradle/Groovy/Junit -- not sure what causes it
but it doesn't need to be an inner class anyway.