Re-work DefaultPublishPackagesExtension for compatibility with Gradle 1.5+.
Since Gradle 1.5 the ivy-publish plugin doesn't let you modify publications,
related tasks, etc. once the PublishingExtension is accessed through the
"project.publishing" property.  Instead you should do all configuration in
"project.publishing {}" blocks.  So, I've re-worked the
DefaultPublishPackagesExtension class to do this.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Don't access the publishing extension, and so don't pass it to DPPE.

* intrepid-plugin\src\main\groovy\holygradle\publishing\DefaultPublishPackagesExtension.groovy
** Don't bother tracking Configurations and PackageArtifactHandlers being
removed, because we only use them to make the default publication within the
project.publishing closure and, by the time that's executed, the configurations
and packaged artifact settings should all be fixed.
** Since Gradle 1.5, the ivy-publish plugin doesn't automatically create a
publication with everything, so we have to explicitly add configurations and
artifacts.
** Also move the task dependency setup for the repbulishing task into the
project.publishing block.

* intrepid-plugin\src\main\groovy\holygradle\publishing\PublishPackagesExtension.groovy
** Instead of providing access to the default publication through a property,
provide a closure to configure it (so that we don't have to access
"project.publishing" as a property).