Fix a deprecation warning and thereby a bug.
I was getting deprecation warnings about modifying a resolution strategy after
the configuration was removed.  That is scheduled to be an error from Gradle 3.0.
It turned out this was due to recently-introduced code in holy-gradle-init.gradle
which was trying to force the "plugin snapshot override" to apply to all Holy
Gradle plugins in all subprojects.  However, that never really made sense
because that code doesn't execute until the subproject buildscript configurations
have been resolved.

Instead, I now just output a warning if the selected plugin versions don't match
the requested ones.  I also added an integration test for that warning.

* .idea\modules\integration\integration_test.iml
* integration\build.gradle
** Added dependency so I can use FileHelper.

* custom-gradle-core-plugin\src\init.d\holy-gradle-init.gradle
** Replace resolutionStrategy with warning.

* custom-gradle-core-plugin\src\test\groovy\holygradle\test\AbstractHolyGradleIntegrationTest.groovy
* integration\src\test\groovy\holygradle\BasicIntegrationTest.groovy
** Integration test for the new warning.