GR #6714 Problems with git credential management
* custom-gradle-core-plugin\src\main\groovy\holygradle\process\ExecHelper.groovy
** This code should have been using ExecResult#assertNormalExitValue (which
throws if the exit code was non-zero; typically used when
ExecSpec#setIgnoreExitValue has been called with true) instead of
ExecResult#rethrowFailure (which is really only intended for internal use by
Gradle's code for running Exec tasks).  Not doing this meant that failed
checkouts were not spotted by the enclosing task, so it could end up retrying
forever.  This code is only used by the GitDependency class, so we never noticed
it when doing Hg checkouts.
** Also renamed the argument throwOnError to throwForExitValue to make more sense.

* intrepid-plugin\src\main\groovy\holygradle\scm\CommandLine.groovy
** Renamed the argument throwOnError to throwForExitValue to make more sense.

* intrepid-plugin\src\main\groovy\holygradle\scm\GitDependency.groovy
** Fixed some error message text which was talking about Mercurial things
instead of Git things.

* intrepid-plugin\src\main\groovy\holygradle\scm\HgDependency.groovy
** Formatting fix, as in GitDependency.

* intrepid-plugin\src\main\groovy\holygradle\source_dependencies\FetchSourceDependencyTask.groovy
** Added "upToDateWhen { false }", otherwise this task's state would be cached
as UP-TO-DATE if it failed.

* custom-gradle-core-plugin\src\test\groovy\holygradle\process\ExecHelperTest.groovy
** New test to make sure I was getting the ExecHelper logic right.