GR #6714 Problems with git credential management
Changes to work with all of wincred, winstore, and manager Git credential helpers.

This code has special handling for wincred (which doesn't pop up a dialog and
just fails if credentials are not available) versus the others (which pop up a
dialog which won't time out on an autobuild machine, so we always pre-emptively
write credentials into the Windows Credential Store, even if they're not needed).
I'd like to simplify it somehow.

* custom-gradle-core-plugin\src\main\groovy\holygradle\process\ExecHelper.groovy
* custom-gradle-core-plugin\src\test\groovy\holygradle\process\ExecHelperTest.groovy
* intrepid-plugin\src\main\groovy\holygradle\scm\CommandLine.groovy
* intrepid-plugin\src\main\groovy\holygradle\scm\SourceControlRepositories.groovy
* intrepid-plugin\src\main\groovy\holygradle\source_dependencies\SourceDependencyHandler.groovy
** Add debug logging of error output stream contents if exec fails.

* intrepid-plugin\src\main\groovy\holygradle\scm\GitDependency.groovy
** Don't try to force configuration of a particular Git credential helper,
because the user might have it set to something else for specific reasons (and
there are at least three possible values on Windows even for the "usual" reasons).
** Add a bit more logging (which I will probably change to debug later).
** If wincred is the configured credential helper, try to check out first and
only cache credentials if that fails.  If some other helper is configured,
always pre-emptively cache credentials to prevent it popping up a dialog.  If no
helper is configured, just try to clone once.

* intrepid-plugin\src\main\groovy\holygradle\scm\GitRepository.groovy
* intrepid-plugin\src\main\groovy\holygradle\scm\GitHelper.groovy
** Factor out code to get a config value, because it needs special handling for
exit value 1, and this is now used in two places.