GR #6715 Allow updating groups of credentials separately
There's a new requirement to allow different credentials to be used for source
depedencies which are hosted at different sites.

I re-wrote credential-store.exe so that it supports association of credentials
with a "basis" credential in a "credential-bases.txt" file.  The Holy Gradle's
intrepid-plugin will add to or update the associations in that file every time
it chooses to cache a credential.  The EXE still allows for crendentials which
are not associated with any credential and provides a separate command for
updating all of those with a single username & password.  I've also added
commands to list (a) credentials associated with a particular basis; and (b)
credentials not associated with any basis.

* credential-store/credential-store.cpp
** Store credentials at "Entreprise" scope instead of "Local Machine", so
they're available if you log on to another machine in the same domain, because
I noticed that's what TortoiseHg does now and it seems more useful.
** Use secure versions of string manipulation functions (sprintf_s, wcsnicmp etc.)
** When looking for credentials with no basis, only return those which match the
given username.  This is because it's possible that a Windows user has
credentials using different (non-Windows) usernames in different locations which
are nothing to do with the Holy Gradle, and we don't want to mess with those.
** Update matching of Git credentials to allow the format used by the two newer
Git credential helpers.
** Don't automatically delete the "Mercurial" credential, because ... I have no
idea why we ever did that.

* credential-store/credential-store.vcxproj
** When I was prompted by VS2013 to stop using insecure string functions, I
thought I might as well bump up the warning level, too.