Fix init-script plugin version checks, and tests thereof.
The init script doesn't need to be checking that all buildscript classpath
modules (in particular, the holygradle plugins) got the requested version,
because that isn't the problem which originally occurred.  The original problem
was that setting any of the main holygradle plugins to a snapshot version
wouldn't override the version of custom-gradle-core-plugin.  So, I've instead
put in a check that no snapshot version of a holygradle plugin ends up depending
on a non-snapshot version.  This doesn't fail the build right now but it needs a
test.  I also think it's slowed down the runtime a lot, so I might just take it
out given that, if the bug ever occurs, it only needs to get fixed once.  Or,
possibly I can just force any holygradle modules which are depended on by a
snapshot version to be a snapshot version (and a changing module) in a
"beforeResolve" callback?

Also, I added a so far empty function to check that the same plugin version is
applied in all projects in a multi-project build, because not dong that is
likely to cause problems.  I haven't written it yet, though.