GR #6279
* intrepid-plugin/src/main/groovy/holygradle/packaging/PackageArtifactBuildScriptHandler.groovy
** When creating a build script, don't create a packed dependency for a source
dependenciy if the source doesn't contain a Gradle project and/or has no
configuration mappings.

* intrepid-plugin/src/main/groovy/holygradle/publishing/DefaultPublishPackagesExtension.groovy
** Don't try to modify any information in the ivy.xml for source dependencies
which aren't Gradle projects and/or to which there are no configuration mappings,
because there will be no relevant dependencies to modify, and indeed its
SourceDependencyHandler intentionally won't provide a module version ID for it.

* intrepid-plugin/src/main/groovy/holygradle/source_dependencies/SourceDependencyHandler.groovy
** Better handling for a case which is probably user error -- if they declare a
source dependency, and the dependency source contains a "build.gradle" (and so
is included in the multi-project build) but there is no configuration mapping
defined (so, there's no proper dependency).  This will now issue a warning.