First attempt at converting the devenv plugin to Kotlin.
It builds, but the tests fail with mysterious "no class def found" exceptions
for Gradle internal classes.

I had to refactor the prerequisites code in custom-gradle-core quite a bit,
with a breaking change to the DSL, otherwise passing in Kotlin lambdas where
closures are needed in Gradle APIs was going to be a pain.