GR #6716 for-defaults now only updates default (non-basis) credentials.
Previously I'd left it calling a reworked version of UpdateAllCredentials, but I
never want to update *all* credentials now, so I just deleted that function and
changed UpdateCredentialsFromDefault to do the right thing.

I added a new command, list-basis, to show the credentials for a given basis.

I allowed list-defaults to take an optional username, since for-defaults
prompts for a username and uses it to find credentials.

I made for-basis exit early with an error if there is no such basis, rather than
prompt you for username & password and only fail after that.