VS2017 hopefully-final code changes.
* credential-store\build.gradle
** More-or-less copy changes from DevEnvHandler.groovy.  The differences are
(a) that vswhereLocation is not stored in a "member" variable; and (b) the final
devenv location is not stored in a top-level property (because we only need it
once per build).

* credential-store\credential-store.cpp
** Fix compile error I introduced in my last commit.

* developer-readme.md
** Note that people need to add chosenDevEnvVersion to gradle.properties.

* devenv-plugin\src\main\groovy\holygradle\devenv\DevEnvHandler.groovy
** Update error message shown when devEnvVersion is not set.
** Pull the rootProject property which caches the devenv location up so that
it's used even if we find devenv.com via an environment variable.
** Construct nextDecimalVersionNumber by string concat.
** Change logic to use "vswhere or env var" to get the common tools dir, then
just derive the devenv.com location from that in one place in the code.
** Use find/collect to find either the 64-bit or 32-bit version of vswhere, for
Groovy style points :-)
** Rename getDevEnvPathFromRegistry to point out that it's not from the registry
but from an environment variable.

* doc\website\sitemap-content.asciidoc
** Update release version in sitemap.