GR #6279 Work around weird Gradle bug related to buildSrc.
After attempting a merge into the "GR_6279_source_override" branch I found that
the new ExecHelper class in buildSrc was being picked up when running tests,
causing the tests to fail because the method overloads are slightly different.

This shouldn't happen in Gradle -- buildSrc classes shouldn't be accessible to
the "real" project's classes -- so I think it's a Gradle bug.  I've worked
around this for now by moving the buildSrc classes into a "holygradle.buildSrc"
namespace.