VS2017 Improvements to finding devenv versions
If the given version is "VSnnn", try the "VSnnnCOMTOOLS" before looking for
vswhere.exe, because some older machines may not have VS2017+ installed and
some versions of vswhere can't find older versions of Visual Studio.

If the given version is "nn.n", use vswhere to find an exact version match.

If the given version is "nn.+", use vswhere to match the major version but
ignore the minor version.

* .idea\inspectionProfiles\Project_Default.xml
** In passing, make "unnecessary 'public' modifier" a very weak warning.

* custom-gradle-core-plugin\custom-gradle-core-plugin.iml
* devenv-plugin\devenv-plugin.iml
* intrepid-plugin\intrepid-plugin.iml
** Newer version of IntelliJ wants to explicitly add Gradle version info.

* devenv-plugin\src\main\groovy\holygradle\devenv\DevEnvHandler.groovy
** Match either "VSnnn", "nn.n", or "nn.+", as described above.
** Make some methods private, in passing.

* doc\website\plugin-devenv.asciidoc
** Update doc for new behaviour.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModuleVersion.groovy
** In passing, fix comment formatting.