GR #6725 Intermittent failure to create link to unpackCache
Certain users were experiencing an intermittent problem where the Holy Gradle
would, instead of correctly rebuilding symlinks somehow leave behind an empty
directory at a link source location, and then fail to re-create either directory
junctions or symlinks (to the unpack cache).

This may be due to some background process keeping a lock on some file
temporarily, we're not sure.  To try to work around this problem until we find
the cause, and because it's in any case not unreasonable, I've modified the link
creation code as follows.  (To understand these explanations, you should know
that the process for creating a directory junction actually involves creating an
empty folder, then creating a "mount point" inside that folder; and likewise in
reverse when deleting.)

* When trying to delete a link (including prior to rebuilding), if there's an
existing empty folder at the source location, it will be deleted and replaced
with a link, in case it was a left-over, half-deleted directory junction.
Previously the Holy Gradle would just print an error and refuse to create the
link.  This runs the risk of deleting an empty directory which the user actually
wanted, but I think that's very unlikely in practice.

* If creation or deletion of the mount point fails, the code will now delete the
folder where it was to be created/deleted, provided that folder is empty.

* If creation or deletion of a symlink or directory junction fails, it is
re-tried up to 10 times, 1 second apart.

* In passing, modified symlink code to use the canonical path (with relative
segments "collapsed") in error/debug messages, as in the junction code.

Specific changes are as follow.

* custom-gradle-core-plugin\src\main\groovy\holygradle\io\FileHelper.groovy
** Factor out helper method.

* custom-gradle-core-plugin\src\main\groovy\holygradle\io\Junction.groovy
** Allow delete/rebuild when there's an existing empty folder at the source.
** Delete empty folder at the source location, if mount point delete/rebuild fails.

* custom-gradle-core-plugin\src\main\groovy\holygradle\io\Link.groovy
** Retry on failure.
** Log what link is being deleted/created, at debug level.

* custom-gradle-core-plugin\src\main\groovy\holygradle\io\Symlink.groovy
** Allow delete/rebuild when there's an existing empty folder at the source.

* custom-gradle-core-plugin\src\test\groovy\holygradle\io\LinkTest.groovy
** In passing, delete links after the tests are done, because occasionally they
end up broken, which causes the next build of the Holy Gradle to fail.

* custom-gradle-core-plugin\src\test\groovy\holygradle\io\LinkTestBase.groovy
** Add tests for new behaviour.