Initial conversion of custom-gradle-core-plugin to Kotlin.
The only intentional substantive change is to strip the generic parameter off
PrerequisitesChecker, because Kotlin revealed that it would really have been
type-erased.  The "parameter" is now always just treated as an array of
Any/Object, with a helper method to provide it as a list of some other type.