GR #6674 Hopefully allow source info to be gathered for project in Hg repos which aren't in the root folder.
This includes coping with Hg repos which have no default remote (typically
because they've never been pushed).  That change was also needed in the the
top-level versionInfo.gradle for building where the "local" Holy Gradle repo is
a dummy, un-pushed repo.

I can't test this until I'm in a position to publish a temp version of the
plugins, though.