GR #6674 Change svn test so it doesn't use a pre-canned working copy.
In order to finish the tests for this feature, I need to add a sub-dir to the
repo and have that count as "not changed" -- that is, I need to commit the
change.  For that to work in svn, I need to be able to commit to the repo, so I
need to check out from and commit to a "file:" URL, instead of using a
pre-canned ZIP of a working copy.