GR #6674 Make sure all unit and integration tests pass
* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\BuildHelper.groovy
** If the build fails for some other reason than a version conflict, the message
of the exception may be null (e.g., for NullPointerException).

* custom-gradle-core-plugin\src\main\groovy\holygradle\io\FileHelper.groovy
** To avoid occasional failures in tests (and elsewhere), try harder to ensure a
file or diretory is really deleted before we return from the delete helper method.

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactBuildScriptHandler.groovy
* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactHandler.groovy
* intrepid-plugin\src\main\groovy\holygradle\scm\SourceControlRepositories.groovy
* intrepid-plugin\src\test\groovy\holygradle\scm\SourceControlRepositoriesTest.groovy
** Move away from having a "null" SourceControlRepository mean something subtly
differen from a "DumySourceControl", because it was causing a
NullPointerException in certain cases.  We always have an object now, and the
dummy means the project/handler folder is not under source control.

* intrepid-plugin\src\main\groovy\holygradle\scm\SvnRepository.groovy
** Fix check for ignored files/folders -- "svn status" only reports ignored for
files or folders which are explicitly ignored, not for files somewhere inside
ignored folders, so we have to walk up the folder hierarchy if we get no status
info for a file.