Fix the last broken integration tests for intrepid!
* intrepid-plugin\src\main\kotlin\holygradle\packaging\PackageArtifactBuildScriptHandler.kt
* intrepid-plugin\src\main\kotlin\holygradle\packaging\PackageArtifactTextFileCollector.kt
** PackageArtifactBuildScriptHandler also needs to be create via the project's
ObjectFactory so that Grrovy Closures passed to the addRepublishing method get
automagically wrapped as Action<>, so that the call reaches Kotlin code.

* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest.groovy
* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest_tBSR_explicitFile.ok.txt
** The explicitFile case has an empty includeBuildScript block and so should
always (or, for a long time) have been producing no build.gradle and
settings.gradle.  I'm not sure why the ".ok" file still said that it should,
but I've just updated it to be empty, and updated the test to expect no settings
file.