Much of SourceOverridesDependencyResolutionListener is unneeded, and some of UnpackModuleVersion.
While reviewing code for the Kotlin conversion, I realised half the code in this
class became obsolete when we changed from having a separate "all-dependencies.xml"
to putting the flattened dependencies of the overriding project into an ivy.xml
file, which allows Gradle to do all the dependency checking for us.  Also some
test files are obsolete.

The doc still needs to be updated, though.

Some methods in UnpackModuleVersion were unused, too.