Only one last integration test to get working, I think.
* integration/src/test/groovy/holygradle/BasicIntegrationTest*
** Tweaks to get the top-level integration regression tests to (nearly) pass.
** I had to change how the republish task was created, and now it doesn't get
created correctly at all, I think, so I still need to fix that.

* intrepid-plugin/src/main/kotlin/holygradle/publishing/DefaultPublishPackagesExtension.kt
** Got this working the way I originally intended.  It was failing after I tried
to fix it because I was logging diagnostic info and calling "asString" instead
of "toString" on the XmlProvider, which invalidates the Node I originally got
from calling asNode -- doh!

TODO: Make sure I'm doing the right @Inject magic for every class which now
specifies methods with Action<> parameters which are being passed Closures from
build scripts.  (That might be the problem with the "republish" task, in fact.)