Pseudo-merge changes on default up to 54863cac9e5d56ac79abd996b98e56033d497f7a
* gradle/wrapper/gradle-wrapper.properties
** Upgrade to Gradle 4.9, to match.

* 48a0dd8dac1d:
** custom-gradle-core-plugin/src/test/groovy/holygradle/test/AbstractHolyGradleIntegrationTest.groovy
*** Comment-back-in the setup to show full stack traces from exceptions in builds.
(I didn't need to make the other changes in this file.)

* 48a0dd8dac1d:
** integration/src/test/groovy/holygradle/BasicIntegrationTest/tAPIT/settings.gradle
** integration/src/test/groovy/holygradle/BasicIntegrationTest/tBPC/settings.gradle
*** Add +enableFeaturePreview('STABLE_PUBLISHING'), now I'm upgrading to Gradle 4.9.

* 8687a7d0641f:
** buildSrc/src/main/groovy/holygradle/buildHelpers/XmlDoc.groovy
** buildSrc/src/main/groovy/holygradle/buildHelpers/XmlDocumentSource.groovy
** website.gradle
*** Fix website building, which presumably got broken by the upgrade to Groovy
2.x in a previous Gradle upgrade.

* 4606f8d45d9b:
** intrepid-plugin/src/main/kotlin/holygradle/unpacking/SpeedyUnpackManyTask.kt
*** Fix dependencies always extracting.  The info file listing what had been
unpacked was always being deleted.  This was supposed to allow you do re-unpack
all and only the required ZIP files if you switched to a different project on
the same machine which used a different subset of configurations for some
unpack module.  However (a) it only deleted the info file, not the unzipped
files, so it never actually did anything useful; and (b) after source override
refactors (I think) it happened much more often, so basically every run of
fetchAllDependencies would unzip everything.

* d87fd76c0109:
** No changes; that revision was just adding a new tag.

* 7f7c94e318f9:
* 3a4285e779b5:
** Fix for gw.bat failing to redirect when there are spaces in the path

* 54863cac9e5d:
** No changes need ported -- this revision was a tidy-up ported *from* the
Kotlin_Conversion branch, to aid later merging.