Fix top-level integration test regression comparison (I hope).
First, I mis-ported the changes from b482b78d01d1 in that I didn't pass the
stripEmptyLines argument as true, and didn't strip "> Configure project :".

Second, I added a change to strip the Holy Gradle source checkout path from the
beginning of a filename in the regression output, but I didn't strip the name of
the checkout folder itself, so the comparison failed if you ran it in a checkout
to a folder with a name other than "holy_gradle_plugins" -- oops.